# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lecturer Contract Generation System API",
    "description": "API documentation for Lecturer Contract Generation System: authentication, recruitment candidates, interview questions, lecturers, courses, classes, course mappings, onboarding, catalog, dashboard, profiles, and users.\n\nSecurity: Most endpoints require a JWT cookie (httpOnly) set by /auth/login. Alternatively you may send an Authorization: Bearer <token> header.\n\nConventions:\n- Pagination: query params page (1-based), limit (default 10, max 100). Responses include page, limit, total, totalPages, hasMore.\n- Sorting: where implemented, optional sort / order query params (not all endpoints expose yet).\n- Filtering: resource-specific query params (e.g., department scoping enforced server-side).\n- Date/time fields are ISO8601 strings in UTC.\n- Errors follow the Error schema.",
    "version": "1.0.0"
  },
...
  "servers": [{ "url": "http://localhost:4000", "description": "Local" }],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "jwt" },
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["superadmin", "admin", "management", "lecturer"] },
          "fullName": { "type": "string", "nullable": true },
          "department": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "lastLogin": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" },
          "code": { "type": "string", "nullable": true },
          "details": { "type": "object", "nullable": true }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "message": "Not authorized" }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "message": "Access denied" }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "message": "Resource not found" }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "message": "Invalid input" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "message": "Email already exists" }
          }
        }
      },
      "ServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "message": "Server error" }
          }
        }
      }
    }
  },
  "security": [{ "cookieAuth": [], "bearerAuth": [] }],
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users",
        "responses": {
          "200": {
            "description": "Users",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        },
        "security": [{ "cookieAuth": [], "bearerAuth": [] }]
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create user",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/ServerError" }
        },
        "security": [{ "cookieAuth": [], "bearerAuth": [] }]
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": ["Users"],
        "summary": "Update user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/ServerError" }
        },
        "security": [{ "cookieAuth": [], "bearerAuth": [] }]
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "message": { "type": "string" } } },
                "example": { "message": "User deleted" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        },
        "security": [{ "cookieAuth": [], "bearerAuth": [] }]
      }
    },
...
  }
}